♻️ take put forksを別ファイルに分けた&eatの処理を別関数に分けた
