+++ philo/Makefile
@@ -6,7 +6,7 @@
 #    By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/08 01:21:55 by miyuu             #+#    #+#              #
-#    Updated: 2025/04/21 21:33:07 by miyuu            ###   ########.fr        #
+#    Updated: 2025/04/21 21:35:59 by miyuu            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,7 +31,8 @@ SRC_FILES = main.c \
 			judgement_stop_thread.c \
 			ft_atoi.c \
 			ft_isdigit.c \
-			utils.c
+			utils.c \
+			take_put_forks.c
 
 # Header files
 HEADER = $(HEADER_DIR)/philo.h
