+++ philo/src/main.c
@@ -6,46 +6,11 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 19:57:59 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/21 21:32:11 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/21 22:13:46 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <philo.h>
-
-void	put_forks(t_thread_arg *philo)
-{
-	pthread_mutex_unlock(philo->first_fork);
-	pthread_mutex_unlock(philo->second_fork);
-}
-
-void	take_forks(t_thread_arg *philo, t_univ_rules rules)
-{
-	if (*philo->can_stop_thread)
-		return ;
-	// printf("%d: philo->first_fork = %d待ち\n", philo->philo_id + 1, philo->first_fork_n);
-	pthread_mutex_lock(philo->first_fork);
-	if (*philo->can_stop_thread)
-	{
-		put_forks(philo);
-		return ;
-	}
-	printf_philo_status("has taken a fork", *philo->start_tv_ms, philo->philo_id + 1, 0);
-	if (rules.total_philo == 1)
-	{
-		safe_usleep(rules.time_die_ms);
-		pthread_mutex_unlock(philo->first_fork);
-		return ;
-	}
-
-	// printf("%d: philo->second_fork = %d待ち\n", philo->philo_id + 1, philo->second_fork_n);
-	pthread_mutex_lock(philo->second_fork);
-	if (*philo->can_stop_thread)
-	{
-		put_forks(philo);
-		return ;
-	}
-	printf_philo_status("has taken a fork", *philo->start_tv_ms, philo->philo_id + 1, 0);
-}
 
 void	thinking_lag(t_univ_rules rules)
 {
@@ -61,6 +26,26 @@ void	thinking_lag(t_univ_rules rules)
 	usleep(think_ms * 0.3 * UNIT_CONV);
 }
 
+int	eat_action(t_thread_arg *data, t_univ_rules rules, int *eat_num)
+{
+	take_forks(data, rules);
+	if (*data->can_stop_thread)
+	{
+		put_forks(data);
+		return (-1);
+	}
+	printf_philo_status("is eating", *data->start_tv_ms, \
+								data->philo_id + 1, 0);
+	*data->last_eat_time = get_now_time_ms();
+	safe_usleep(rules.time_eat_ms);
+	put_forks(data);
+	if (*data->can_stop_thread)
+		return (-1);
+	if (rules.must_eat != -1 && ++(*eat_num) >= rules.must_eat)
+		*data->is_eat_full = true;
+	return (0);
+}
+
 void	*action_philosophers(void *arg)
 {
 	t_thread_arg	*data;
@@ -73,27 +58,13 @@ void	*action_philosophers(void *arg)
 	eat_num = 0;
 	while (!*data->can_start_eat)
 		;
-	// printf("開始!%d, time: %ld\n", data->philo_id + 1, get_now_time_ms());
 	if (data->philo_id % 2 == 0)
 		thinking_lag(rules);
 	while (!*data->can_stop_thread)
 	{
-		take_forks(data, rules);
-		if (*data->can_stop_thread)
-		{
-			put_forks(data);
+		if (eat_action(data, rules, &eat_num) != 0)
 			break ;
-		}
-		/* eatを開始 */
-		now_ms = printf_philo_status("is eating", *data->start_tv_ms, data->philo_id + 1, now_ms);
-		*data->last_eat_time = now_ms;
-		safe_usleep(rules.time_eat_ms);
-		put_forks(data);
 
-		if (*data->can_stop_thread)
-			break ;
-		if (++eat_num >= rules.must_eat && rules.must_eat != -1)
-			*data->is_eat_full = true;
 		if (*data->can_stop_thread)
 			break ;
 		/* eatが終わり、sleepを開始 */
